reverting: experiment failed
the `perl:` section experiment (per #1) failed.  Moved to [https://ci.appveyor.com/tools/validate-yaml](https://ci.appveyor.com/tools/validate-yaml) for on-the-fly testing of the YML, and it told me that `perl:` was an `invalid setting or section`.  I also tried an `environment: perl: default`, but that would only allow one.  I tried setting that per-configuration, but that didn't work, either.

I may come back to this, but for now, I don't think it's going to work as `%perl%`.

Maybe I'll just set the `%perl%` variable based on `%configuration%` during the `install:`...